+++ examples/global.js
@@ -1,0 +1,12 @@
+(window.foremarkConfig = window.foremarkConfig || []).push({
+    'sitemap.merge': {
+        'main': [
+            ['Foremark Manual', [], [
+                ['The Foremark Book', ['/book.mf.xhtml']],
+                ['Foremark Reference Manual', ['/reference.mf.xhtml']],
+                ['Tools for writing Foremark', ['/tools.mf.xhtml']],
+            ]],
+            ['Media element test', ['/media.mf.xhtml']],
+        ],
+    },
+});