+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added the WOFF2 version of every included web font. This reduces the downloaded asset size by up to 160KB.
 - Changed the way viewer configuration objects are supplied. Configuration objects now can be provided from asynchronously loaded script tags, paving the way to collectively customize the behavior of a set of documents.
 - Fixed an issue that the selection of the table of contents does not follow the scroll position.
+- A sitemap can be defined in configuration. The sitemap is displayed in the table of contents together with a current page's headings.
 
 ## [0.0.10] - 2018-01-05